pip-requirements.txt
albumentations==0.4.3
geopandas>=0.7.0
ipykernel
ipywidgets
numba
opencv-python>=4.1
osmnx
p_tqdm
statsmodels
torch>=1.3.1
torchsummary
torchvision
utm
